Recovered and fixed lost files issue
