+++ src/api/docker-compose.yml
@@ -82,3 +82,61 @@ services:
       - 'traefik.http.middlewares.strip_auth_prefix.stripprefix.prefixes=/${API_VERSION}/auth'
       - 'traefik.http.middlewares.strip_auth_prefix.stripprefix.forceSlash=true'
       - 'traefik.http.routers.auth.middlewares=strip_auth_prefix'
+
+  redis:
+    image: redis:latest
+    container_name: api_redis
+    ports:
+      - '6379'
+    command: ['redis-server', '--appendonly', 'yes']
+    volumes:
+      - ./redis-data:/data
+
+  # posts service
+  posts:
+    container_name: 'posts'
+    build:
+      context: ./posts
+      dockerfile: Dockerfile
+    environment:
+      - POSTS_PORT
+    ports:
+      - ${POSTS_PORT}
+    depends_on:
+      - traefik
+      - redis
+    labels:
+      # Enable Traefik
+      - 'traefik.enable=true'
+      # Traefik routing for the image service at /v1/image
+      - 'traefik.http.routers.posts.rule=Host(`${API_HOST}`) && PathPrefix(`/${API_VERSION}/posts`)'
+      # Specify the posts service port
+      - 'traefik.http.services.posts.loadbalancer.server.port=${IMAGE_PORT}'
+      # Add middleware to this route to strip the /v1/posts prefix
+      - 'traefik.http.middlewares.strip_posts_prefix.stripprefix.prefixes=/${API_VERSION}/posts'
+      - 'traefik.http.middlewares.strip_posts_prefix.stripprefix.forceSlash=true'
+      - 'traefik.http.routers.posts.middlewares=strip_posts_prefix'
+
+  # SSO Identity Provider test service, https://simplesamlphp.org
+  # Access to the login page available at http://login.localhost/simplesaml
+  login:
+    image: kristophjunge/test-saml-idp
+    container_name: 'login'
+    ports:
+      - '8080'
+    depends_on:
+      - traefik
+    environment:
+      - SIMPLESAMLPHP_SP_ENTITY_ID=${SAML_ENTITY_ID}
+      - SIMPLESAMLPHP_SP_ASSERTION_CONSUMER_SERVICE=${SSO_LOGIN_CALLBACK_URL}
+      - SIMPLESAMLPHP_SP_SINGLE_LOGOUT_SERVICE=${SLO_LOGOUT_CALLBACK_URL}
+    volumes:
+      - ./auth/config/simplesamlphp-users.php:/var/www/simplesamlphp/config/authsources.php
+      - ./auth/config/saml20-idp-hosted.php:/var/www/simplesamlphp/metadata/saml20-idp-hosted.php
+    labels:
+      # Enable Traefik
+      - 'traefik.enable=true'
+      # Traefik routing for login at http://login.localhost/simplesaml
+      - 'traefik.http.routers.login.rule=Host(`login.localhost`)'
+      # Specify the login port
+      - 'traefik.http.services.login.loadbalancer.server.port=8080'
