+++ src/api/posts/jest.config.js
@@ -3,6 +3,6 @@ const baseConfig = require('../../../jest.config.base');
 module.exports = {
   ...baseConfig,
   rootDir: '../../..',
-  testMatch: ['<rootDir>/src/api/posts/tests/*.test.js'],
+  testMatch: ['<rootDir>/src/api/posts/test/*.test.js'],
   collectCoverageFrom: ['<rootDir>/src.api/posts/src/**/*.js'],
 };
