--- src/api/posts/src/bin/article-error.js
@@ -1,13 +1,0 @@
-class ArticleError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = this.constructor.name;
-    if (typeof Error.captureStackTrace === 'function') {
-      Error.captureStackTrace(this, this.constructor);
-    } else {
-      this.stack = new Error(message).stack;
-    }
-  }
-}
-
-module.exports = ArticleError;
