--- src/api/posts/src/bin/processHTML/remove-empty-anchor.js
@@ -1,12 +1,0 @@
-module.exports = function (dom) {
-  if (!(dom && dom.window && dom.window.document)) {
-    return;
-  }
-
-  dom.window.document.querySelectorAll('a').forEach((a) => {
-    const anchorInnerHTML = a.innerHTML;
-    if (!anchorInnerHTML.trim()) {
-      a.remove();
-    }
-  });
-};
