--- src/api/posts/src/bin/processHTML/syntax-highlight.js
@@ -1,60 +1,0 @@
-const hljs = require('highlight.js/lib/core');
-
-// Tweak the language list here, see https://highlightjs.org/usage/
-hljs.registerLanguage('cLike', require('highlight.js/lib/languages/c-like'));
-hljs.registerLanguage('sql', require('highlight.js/lib/languages/sql'));
-hljs.registerLanguage('powershell', require('highlight.js/lib/languages/powershell'));
-hljs.registerLanguage('csharp', require('highlight.js/lib/languages/csharp'));
-hljs.registerLanguage('go', require('highlight.js/lib/languages/go'));
-hljs.registerLanguage('css', require('highlight.js/lib/languages/css'));
-hljs.registerLanguage('makefile', require('highlight.js/lib/languages/makefile'));
-hljs.registerLanguage('markdown', require('highlight.js/lib/languages/markdown'));
-hljs.registerLanguage('swift', require('highlight.js/lib/languages/swift'));
-hljs.registerLanguage('armasm', require('highlight.js/lib/languages/armasm'));
-hljs.registerLanguage('diff', require('highlight.js/lib/languages/diff'));
-hljs.registerLanguage('python', require('highlight.js/lib/languages/python'));
-hljs.registerLanguage('dockerfile', require('highlight.js/lib/languages/dockerfile'));
-hljs.registerLanguage('http', require('highlight.js/lib/languages/http'));
-hljs.registerLanguage('typescript', require('highlight.js/lib/languages/typescript'));
-hljs.registerLanguage('ini', require('highlight.js/lib/languages/ini'));
-hljs.registerLanguage('bash', require('highlight.js/lib/languages/bash'));
-hljs.registerLanguage('nginx', require('highlight.js/lib/languages/nginx'));
-hljs.registerLanguage('java', require('highlight.js/lib/languages/java'));
-hljs.registerLanguage('javascript', require('highlight.js/lib/languages/javascript'));
-hljs.registerLanguage('rust', require('highlight.js/lib/languages/rust'));
-hljs.registerLanguage('json', require('highlight.js/lib/languages/json'));
-hljs.registerLanguage('x86asm', require('highlight.js/lib/languages/x86asm'));
-hljs.registerLanguage('kotlin', require('highlight.js/lib/languages/kotlin'));
-hljs.registerLanguage('xml', require('highlight.js/lib/languages/xml'));
-hljs.registerLanguage('shell', require('highlight.js/lib/languages/shell'));
-hljs.registerLanguage('yaml', require('highlight.js/lib/languages/yaml'));
-hljs.registerLanguage('php', require('highlight.js/lib/languages/php'));
-hljs.registerLanguage('dos', require('highlight.js/lib/languages/dos'));
-
-/**
- * Given a parsed JSDOM Object, find all <pre>...</pre> code blocks
- * and apply syntax highlight markup.
- */
-module.exports = function (dom) {
-  if (!(dom && dom.window && dom.window.document)) {
-    return;
-  }
-
-  dom.window.document.querySelectorAll('pre code').forEach((code) => {
-    const { value, language } = hljs.highlightAuto(code.innerHTML);
-
-    // If the language wasn't detected, return the HTML untouched
-    if (!language) {
-      return;
-    }
-
-    // Otherwise, decorate the <pre> with class names for highlighting this language
-    const pre = code.parentNode;
-    if (pre) {
-      pre.classList.add('hljs', language);
-    }
-
-    // Replace the contents with newly marked up syntax highlighting
-    code.innerHTML = value;
-  });
-};
