+++ src/api/posts/src/index.js
@@ -4,6 +4,6 @@ const posts = require('./routes/posts');
 
 const service = new Satellite();
 
-service.router.use('/posts', posts);
+service.router.use('/', posts);
 
 module.exports = service;
