+++ src/api/posts/src/libs/redis.js
@@ -1,10 +1,9 @@
 const Redis = require('ioredis');
 const MockRedis = require('ioredis-mock');
 const { logger } = require('@senecacdot/satellite');
-const parseUrl = require('../bin/url-parser');
 
 // If you need to set the Redis URL, do it in REDIS_URL
-const redisUrl = parseUrl(process.env.REDIS_URL, process.env.REDIS_PORT) || 'redis://redis:6379';
+const redisUrl = 'http://redis:6379';
 
 // Set MOCK_REDIS=1 to mock, MOCK_REDIS= to use real redis
 const useMockRedis = process.env.MOCK_REDIS;
