+++ src/api/posts/src/storage.js
@@ -1,5 +1,5 @@
 const { logger } = require('@senecacdot/satellite');
-const { redis } = require('../libs/redis');
+const { redis } = require('./libs/redis');
 // Redis Keys
 const feedsKey = 't:feeds';
 const flaggedFeedsKey = 't:feeds:flagged';
